worldenv/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-worldenv"
version = "0.1.0"
description = "Sample any RL environment from a General Unified World Model"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
full = [
    "general-unified-world-model>=0.0.3",
    "torch>=2.0",
    "gymnasium>=0.29",
]
dev = ["pytest>=8.0.0"]

[project.urls]
Homepage = "https://github.com/JacobFV/general-unified-world-modeling"
Documentation = "https://jacobfv.github.io/general-unified-world-modeling/environments/"

[project.scripts]
server = "worldenv.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["worldenv", "worldenv.server"]
package-dir = { "worldenv" = ".", "worldenv.server" = "server" }
